requirements.txt
streamlit>=1.26.0
mysql-connector-python>=8.1.0
pandas>=2.1.0
jinja2>=3.1.2
types-Jinja2>=2.11.9
python-dotenv>=1.0.0
cryptography>=41.0.0
